+++ deepchem/models/torch_models/tests/test_olmo_classification.py
@@ -16,21 +16,8 @@ SMILES = [
 ]
 
 
-def quantization_config():
-    if not torch.cuda.is_available():
-        # bitsandbytes 4-bit quantization requires a CUDA GPU.
-        return None
-    from transformers import BitsAndBytesConfig
-    return BitsAndBytesConfig(
-        load_in_4bit=True,
-        bnb_4bit_quant_type="nf4",
-        bnb_4bit_use_double_quant=True,
-        bnb_4bit_compute_dtype=torch.float16,
-    )
-
-
 @pytest.mark.hf
-def test_olmo_single_label_classification():
+def test_olmo_single_label_classification(quantization_config):
     pytest.skip(
         "Skipping slow OLMo test that downloads and runs the full model.")
     from deepchem.models.torch_models.olmo import Olmo
@@ -41,7 +28,7 @@ def test_olmo_single_label_classification():
         n_tasks=1,
         torch_dtype=torch.float16
         if torch.cuda.is_available() else torch.float32,
-        quantization_config=quantization_config(),
+        quantization_config=quantization_config,
         skip_weight_init=True,
         gradient_checkpointing=True,
     )
@@ -59,7 +46,7 @@ def test_olmo_single_label_classification():
 
 
 @pytest.mark.hf
-def test_olmo_single_label_classification_overfit():
+def test_olmo_single_label_classification_overfit(quantization_config):
     pytest.skip(
         "Skipping slow OLMo test that downloads and runs the full model.")
     from deepchem.models.torch_models.olmo import Olmo
@@ -69,7 +56,7 @@ def test_olmo_single_label_classification_overfit():
                  n_tasks=1,
                  torch_dtype=torch.float16
                  if torch.cuda.is_available() else torch.float32,
-                 quantization_config=quantization_config(),
+                 quantization_config=quantization_config,
                  skip_weight_init=True,
                  gradient_checkpointing=True)
     model.load_from_pretrained("allenai/OLMo-1B-hf", from_hf_checkpoint=True)
@@ -173,7 +160,7 @@ def test_olmo_single_label_classification_save_reload(tmpdir):
 
 
 @pytest.mark.hf
-def test_olmo_multi_label_classification():
+def test_olmo_multi_label_classification(quantization_config):
     pytest.skip(
         "Skipping slow OLMo test that downloads and runs the full model.")
     from deepchem.models.torch_models.olmo import Olmo
@@ -183,7 +170,7 @@ def test_olmo_multi_label_classification():
                  n_tasks=2,
                  torch_dtype=torch.float16
                  if torch.cuda.is_available() else torch.float32,
-                 quantization_config=quantization_config(),
+                 quantization_config=quantization_config,
                  skip_weight_init=True,
                  gradient_checkpointing=True)
     model.load_from_pretrained("allenai/OLMo-1B-hf", from_hf_checkpoint=True)
@@ -200,7 +187,7 @@ def test_olmo_multi_label_classification():
 
 
 @pytest.mark.hf
-def test_olmo_multi_label_classification_overfit():
+def test_olmo_multi_label_classification_overfit(quantization_config):
     pytest.skip(
         "Skipping slow OLMo test that downloads and runs the full model.")
     from deepchem.models.torch_models.olmo import Olmo
@@ -210,7 +197,7 @@ def test_olmo_multi_label_classification_overfit():
                  n_tasks=2,
                  torch_dtype=torch.float16
                  if torch.cuda.is_available() else torch.float32,
-                 quantization_config=quantization_config(),
+                 quantization_config=quantization_config,
                  skip_weight_init=True,
                  gradient_checkpointing=True)
     model.load_from_pretrained("allenai/OLMo-1B-hf", from_hf_checkpoint=True)
