+++ deepchem/models/torch_models/tests/test_olmo_regression.py
@@ -16,21 +16,8 @@ SMILES = [
 ]
 
 
-def quantization_config():
-    if not torch.cuda.is_available():
-        # bitsandbytes 4-bit quantization requires a CUDA GPU.
-        return None
-    from transformers import BitsAndBytesConfig
-    return BitsAndBytesConfig(
-        load_in_4bit=True,
-        bnb_4bit_quant_type="nf4",
-        bnb_4bit_use_double_quant=True,
-        bnb_4bit_compute_dtype=torch.float16,
-    )
-
-
 @pytest.mark.hf
-def test_olmo_single_label_regression():
+def test_olmo_single_label_regression(quantization_config):
     pytest.skip(
         "Skipping slow OLMo test that downloads and runs the full model.")
     from deepchem.models.torch_models.olmo import Olmo
@@ -40,7 +27,7 @@ def test_olmo_single_label_regression():
                  n_tasks=1,
                  torch_dtype=torch.float16
                  if torch.cuda.is_available() else torch.float32,
-                 quantization_config=quantization_config(),
+                 quantization_config=quantization_config,
                  skip_weight_init=True,
                  gradient_checkpointing=True)
     model.load_from_pretrained("allenai/OLMo-1B-hf", from_hf_checkpoint=True)
@@ -57,7 +44,7 @@ def test_olmo_single_label_regression():
 
 
 @pytest.mark.hf
-def test_olmo_single_label_regression_overfit():
+def test_olmo_single_label_regression_overfit(quantization_config):
     pytest.skip(
         "Skipping slow OLMo test that downloads and runs the full model.")
     from deepchem.models.torch_models.olmo import Olmo
@@ -67,7 +54,7 @@ def test_olmo_single_label_regression_overfit():
                  n_tasks=1,
                  torch_dtype=torch.float16
                  if torch.cuda.is_available() else torch.float32,
-                 quantization_config=quantization_config(),
+                 quantization_config=quantization_config,
                  skip_weight_init=True,
                  gradient_checkpointing=True)
     model.load_from_pretrained("allenai/OLMo-1B-hf", from_hf_checkpoint=True)
@@ -171,7 +158,7 @@ def test_olmo_single_label_regression_save_reload(tmpdir):
 
 
 @pytest.mark.hf
-def test_olmo_multi_label_regression():
+def test_olmo_multi_label_regression(quantization_config):
     pytest.skip(
         "Skipping slow OLMo test that downloads and runs the full model.")
     from deepchem.models.torch_models.olmo import Olmo
@@ -181,7 +168,7 @@ def test_olmo_multi_label_regression():
                  n_tasks=2,
                  torch_dtype=torch.float16
                  if torch.cuda.is_available() else torch.float32,
-                 quantization_config=quantization_config(),
+                 quantization_config=quantization_config,
                  skip_weight_init=True,
                  gradient_checkpointing=True)
     model.load_from_pretrained("allenai/OLMo-1B-hf", from_hf_checkpoint=True)
@@ -198,7 +185,7 @@ def test_olmo_multi_label_regression():
 
 
 @pytest.mark.hf
-def test_olmo_multi_label_regression_overfit():
+def test_olmo_multi_label_regression_overfit(quantization_config):
     pytest.skip(
         "Skipping slow OLMo test that downloads and runs the full model.")
     from deepchem.models.torch_models.olmo import Olmo
@@ -208,7 +195,7 @@ def test_olmo_multi_label_regression_overfit():
                  n_tasks=2,
                  torch_dtype=torch.float16
                  if torch.cuda.is_available() else torch.float32,
-                 quantization_config=quantization_config(),
+                 quantization_config=quantization_config,
                  skip_weight_init=True,
                  gradient_checkpointing=True)
     model.load_from_pretrained("allenai/OLMo-1B-hf", from_hf_checkpoint=True)
